requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
build==0.8.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
idna==3.4
one-piece-downloader-cli @ file:///Users/rvallejos/Documents/github/CLIs/one-piece-downloader-cli/dist/one_piece_downloader_cli-0.1.0-py3-none-any.whl
packaging==21.3
pep517==0.13.0
Pillow==9.2.0
pyparsing==3.0.9
requests==2.28.1
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.12
